+++ src/server/Server.java
@@ -49,12 +49,12 @@ public class Server {
 		// Push/Send game/player stats to DB/client
 	}
 	*/
-	static int gamesize = 0;
+	static long gamesize = 0;
     static Map<Player, ObjectInputStream> connected_playerInput = null;
     static Map<Player, ObjectOutputStream> connected_playerOutput = null;
     static Map<Player, Thread> connected_playerThread = null;
-	static Map<Integer, GameThread> connected_games = null;
-	static Map<Integer, Thread> connected_gamethreads = null;
+	static Map<Long, GameThread> connected_games = null;
+	static Map<Long, Thread> connected_gamethreads = null;
 	static int portNumber;
 	
     /**
@@ -68,8 +68,8 @@ public class Server {
 			System.err.print("Usage: java Server <port Number>");
 			System.exit(1);
 		}
-		connected_games = Collections.synchronizedMap(new HashMap<Integer, GameThread>());
-		connected_gamethreads = Collections.synchronizedMap(new HashMap<Integer, Thread>());
+		connected_games = Collections.synchronizedMap(new HashMap<Long, GameThread>());
+		connected_gamethreads = Collections.synchronizedMap(new HashMap<Long, Thread>());
 		connected_playerInput = Collections.synchronizedMap(new HashMap<Player,ObjectInputStream>());
 		connected_playerOutput = Collections.synchronizedMap(new HashMap<Player,ObjectOutputStream>());
 		connected_playerThread = Collections.synchronizedMap(new HashMap<Player,Thread>());
